# config/swagger.nodos.json
{
    "openapi": "3.0.0",
    "info": {
       "version": "1.0",
       "title": "API red de sensores",
       "description": "Documentacion de los endpoints disponibles para el proyecto de red de sensores",
       "license": {
          "name": "ITESO",
          "url": "https://www.iteso.mx/"
       }
    },
    "paths": {
...
       "/usuario/edit/user-password/{userId}": {
          "patch": {
             "tags": [
                "Users"
             ],
             "security": [
                "BearerAuth"
             ],
             "summary": "Editar contraseña de cualquier usuario mediante su ID",
             "description": "Para editar la contraseña de cualquier usuario (ADMINS)",
             "operationId": "patchUserPassword",
             "parameters": [
                {
                   "in": "body",
                   "name": "userId",
                   "description": "",
                   "required": true,
                   "schema": {
                      "type": "object",
                      "properties": {
                         "password": {
                            "type": "string"
                         }
                      }
                   }
                },
                {
                   "in": "header",
                   "name": "Authorization",
                   "schema": {
                      "type": "string"
                   },
                   "description": "Bearer token",
                   "required": true
                }
             ],
             "responses": {
                "200": {
                   "description": "Ok"
                },
                "400": {
                   "description": "Error"
                },
                "500": {
                   "description": "Internal error"
                }
             }
          }
       },
       "/usuario/edit/my-password/{userId}": {
         "patch": {
            "tags": [
               "Users"
            ],
            "security": [
               "BearerAuth"
            ],
            "summary": "Editar contraseña del usuario actual mediante su ID",
            "description": "Solo para editar la contraseña del usuario actual con la que se tiene iniciada la sesion",
            "operationId": "patchMyPassword",
            "parameters": [
               {
                  "in": "body",
                  "name": "data",
                  "description": "",
                  "required": true,
                  "schema": {
                     "type": "object",
                     "properties": {
                        "old_password": {
                           "type": "string"
                        },
                        "new_password": {
                           "type": "string"
                        }
                     }
                  }
               },
               {
                  "in": "header",
                  "name": "Authorization",
                  "schema": {
                     "type": "string"
                  },
                  "description": "Bearer token",
                  "required": true
               }
            ],
            "responses": {
               "200": {
                  "description": "Ok"
               },
               "400": {
                  "description": "Error"
               },
               "401": {
                  "description": "Unauthorized"
               },
               "500": {
                  "description": "Internal error"
               }
            }
         }
      },
      "/usuario/edit/name/{userId}": {
        "patch": {
           "tags": [
              "Users"
           ],
           "security": [
              "BearerAuth"
           ],
           "summary": "Editar name del usuario actual mediante su ID",
           "description": "",
           "operationId": "patchName",
           "parameters": [
              {
                 "in": "body",
                 "name": "data",
                 "description": "",
                 "required": true,
                 "schema": {
                    "type": "object",
                    "properties": {
                       "name": {
                          "type": "string"
                       }
                    }
                 }
              },
              {
                 "in": "header",
                 "name": "Authorization",
                 "schema": {
                    "type": "string"
                 },
                 "description": "Bearer token",
                 "required": true
              }
           ],
           "responses": {
              "200": {
                 "description": "Ok"
              },
              "400": {
                 "description": "Error"
              },
              "401": {
                 "description": "Unauthorized"
              },
              "500": {
                 "description": "Internal error"
              }
           }
        }
     },
...
    },
    "servers": [
       {
          "description": "Host",
          "url": "http://papvidadigital-test.com/wsnv21/api"
       }
    ],
    "components": {
       "securitySchemes": {
          "BearerAuth": {
             "type": "http",
             "scheme": "bearer"
          }
       },
...
    }
 }